requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloud-sql-python-connector==1.2.4
cryptography==41.0.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
